requirements.txt
autopep8==1.5.4
boto3==1.15.17
botocore==1.18.17
jmespath==0.10.0
mysql-connector==2.2.9
mysql-connector-python==8.0.21
names==0.3.0
protobuf==3.13.0
pycodestyle==2.6.0
python-dateutil==2.8.1
python-decouple==3.3
s3transfer==0.3.3
six==1.15.0
toml==0.10.1
urllib3==1.25.10
